third_party/blink/web_tests/platform/generic/fast/css/deprecated-flex-box-zero-width-intrinsic-max-width-expected.txt

Tests if the Deprecated Flex boxes with max-width: max-content and width:0 has non-zero calculated width

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS flexBoxDiv.width is non-zero.
PASS successfullyParsed is true

TEST COMPLETE
